Rewrite analysis.py to do everything inside gdb
analysis.py is now used to launch gdb, where it then runs as a gdb script and downloads and analyses crashes from a specified campaign.
This differs from previous versions where it was invoked for the entire mothership (i.e. every campaign).
The new design is significantly faster as there is less overhead of relaunching gdb.